+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Other blocks are:
 - **LIB** - New `xunit` and `yunit` attributes in class `Spectrum`.
 - **BRS** - Support of hybrid schemes and composite analyses mixing different level of calculation.
 - **BLS** - Scaling now supports more operations (see documentation).
+- **API** - Support of electronic energies in Gaussian log.
 
 ### Fixed